Extract snapshot alias lock helpers
