+++ lib/instances/fork.go
@@ -10,7 +10,6 @@ import (
 	"strings"
 	"time"
 
-	"github.com/kernel/hypeman/lib/forkvm"
 	"github.com/kernel/hypeman/lib/guest"
 	"github.com/kernel/hypeman/lib/hypervisor"
 	"github.com/kernel/hypeman/lib/instances/phasetracking"
@@ -259,24 +258,7 @@ func (m *manager) forkInstanceFromStoppedOrStandby(ctx context.Context, id strin
 	})
 	defer cu.Clean()
 
-	if err := func() error {
-		unlockAliasReaders := hypervisor.LockSnapshotSourceAliasReaders()
-		defer unlockAliasReaders()
-
-		if source.State == StateStandby {
-			if err := m.ensureSnapshotMemoryReady(ctx, m.paths.InstanceSnapshotLatest(id), m.snapshotJobKeyForInstance(id), stored.HypervisorType); err != nil {
-				return fmt.Errorf("prepare standby snapshot for fork: %w", err)
-			}
-		}
-
-		if err := forkvm.CopyGuestDirectory(srcDir, dstDir); err != nil {
-			if errors.Is(err, forkvm.ErrSparseCopyUnsupported) {
-				return fmt.Errorf("fork requires sparse-capable filesystem (SEEK_DATA/SEEK_HOLE unsupported): %w", err)
-			}
-			return fmt.Errorf("clone guest directory: %w", err)
-		}
-		return nil
-	}(); err != nil {
+	if err := m.copyForkSourceGuestDirectory(ctx, source.State, id, stored, srcDir, dstDir); err != nil {
 		return nil, err
 	}
 
@@ -338,20 +320,15 @@ func (m *manager) forkInstanceFromStoppedOrStandby(ctx context.Context, id strin
 		if forkMeta.NetworkEnabled {
 			netCfg = &hypervisor.ForkNetworkConfig{TAPDevice: network.GenerateTAPName(forkID)}
 		}
-		err := func() error {
-			unlockAliasReaders := hypervisor.LockSnapshotSourceAliasReaders()
-			defer unlockAliasReaders()
-			_, err := starter.PrepareFork(ctx, hypervisor.ForkPrepareRequest{
-				SnapshotConfigPath: snapshotConfigPath,
-				SourceDataDir:      stored.DataDir,
-				TargetDataDir:      forkMeta.DataDir,
-				VsockCID:           forkMeta.VsockCID,
-				VsockSocket:        forkMeta.VsockSocket,
-				SerialLogPath:      m.paths.InstanceAppLog(forkID),
-				Network:            netCfg,
-			})
-			return err
-		}()
+		err := prepareForkWithAliasReadLock(ctx, starter, hypervisor.ForkPrepareRequest{
+			SnapshotConfigPath: snapshotConfigPath,
+			SourceDataDir:      stored.DataDir,
+			TargetDataDir:      forkMeta.DataDir,
+			VsockCID:           forkMeta.VsockCID,
+			VsockSocket:        forkMeta.VsockSocket,
+			SerialLogPath:      m.paths.InstanceAppLog(forkID),
+			Network:            netCfg,
+		})
 		if err != nil {
 			if errors.Is(err, hypervisor.ErrNotSupported) {
 				return nil, fmt.Errorf("%w: fork is not supported for hypervisor %s", ErrNotSupported, stored.HypervisorType)
