+++ lib/instances/snapshot.go
@@ -412,22 +412,7 @@ func (m *manager) forkSnapshot(ctx context.Context, snapshotID string, req ForkS
 	if target != nil && target.State == compressionJobStateRunning {
 		m.recordSnapshotCompressionPreemption(ctx, snapshotCompressionPreemptionForkSnapshot, target.Target)
 	}
-	if err := func() error {
-		unlockAliasReaders := hypervisor.LockSnapshotSourceAliasReaders()
-		defer unlockAliasReaders()
-
-		if err := m.ensureSnapshotMemoryReady(ctx, m.paths.SnapshotGuestDir(snapshotID), "", rec.StoredMetadata.HypervisorType); err != nil {
-			return fmt.Errorf("prepare snapshot memory for fork: %w", err)
-		}
-
-		if err := forkvm.CopyGuestDirectory(m.paths.SnapshotGuestDir(snapshotID), dstDir); err != nil {
-			if errors.Is(err, forkvm.ErrSparseCopyUnsupported) {
-				return fmt.Errorf("fork from snapshot requires sparse-capable filesystem (SEEK_DATA/SEEK_HOLE unsupported): %w", err)
-			}
-			return fmt.Errorf("clone snapshot payload: %w", err)
-		}
-		return nil
-	}(); err != nil {
+	if err := m.copySnapshotGuestDirectoryForFork(ctx, snapshotID, rec.StoredMetadata.HypervisorType, dstDir); err != nil {
 		return nil, err
 	}
 
@@ -478,20 +463,15 @@ func (m *manager) forkSnapshot(ctx context.Context, snapshotID string, req ForkS
 		if forkMeta.NetworkEnabled {
 			netCfg = &hypervisor.ForkNetworkConfig{TAPDevice: network.GenerateTAPName(forkID)}
 		}
-		err := func() error {
-			unlockAliasReaders := hypervisor.LockSnapshotSourceAliasReaders()
-			defer unlockAliasReaders()
-			_, err := starter.PrepareFork(ctx, hypervisor.ForkPrepareRequest{
-				SnapshotConfigPath: m.paths.InstanceSnapshotConfig(forkID),
-				SourceDataDir:      rec.StoredMetadata.DataDir,
-				TargetDataDir:      forkMeta.DataDir,
-				VsockCID:           forkMeta.VsockCID,
-				VsockSocket:        forkMeta.VsockSocket,
-				SerialLogPath:      m.paths.InstanceAppLog(forkID),
-				Network:            netCfg,
-			})
-			return err
-		}()
+		err := prepareForkWithAliasReadLock(ctx, starter, hypervisor.ForkPrepareRequest{
+			SnapshotConfigPath: m.paths.InstanceSnapshotConfig(forkID),
+			SourceDataDir:      rec.StoredMetadata.DataDir,
+			TargetDataDir:      forkMeta.DataDir,
+			VsockCID:           forkMeta.VsockCID,
+			VsockSocket:        forkMeta.VsockSocket,
+			SerialLogPath:      m.paths.InstanceAppLog(forkID),
+			Network:            netCfg,
+		})
 		if err != nil {
 			if errors.Is(err, hypervisor.ErrNotSupported) {
 				return nil, fmt.Errorf("%w: snapshot fork is not supported for hypervisor %s", ErrNotSupported, targetHypervisor)
@@ -516,11 +496,7 @@ func (m *manager) forkSnapshot(ctx context.Context, snapshotID string, req ForkS
 }
 
 func (m *manager) copySnapshotPayload(sourceInstanceID, snapshotGuestDir string) error {
-	err := func() error {
-		unlockAliasReaders := hypervisor.LockSnapshotSourceAliasReaders()
-		defer unlockAliasReaders()
-		return forkvm.CopyGuestDirectory(m.paths.InstanceDir(sourceInstanceID), snapshotGuestDir)
-	}()
+	err := copyGuestDirectoryWithAliasReadLock(m.paths.InstanceDir(sourceInstanceID), snapshotGuestDir)
 	if err != nil {
 		if errors.Is(err, forkvm.ErrSparseCopyUnsupported) {
 			return fmt.Errorf("snapshot requires sparse-capable filesystem (SEEK_DATA/SEEK_HOLE unsupported): %w", err)
