+++ lib/instances/snapshot_alias_lock.go
@@ -1,0 +1,61 @@
+package instances
+
+import (
+	"context"
+	"errors"
+	"fmt"
+
+	"github.com/kernel/hypeman/lib/forkvm"
+	"github.com/kernel/hypeman/lib/hypervisor"
+)
+
+func withSnapshotSourceAliasReadLock(run func() error) error {
+	unlockAliasReaders := hypervisor.LockSnapshotSourceAliasReaders()
+	defer unlockAliasReaders()
+	return run()
+}
+
+func prepareForkWithAliasReadLock(ctx context.Context, starter hypervisor.VMStarter, req hypervisor.ForkPrepareRequest) error {
+	return withSnapshotSourceAliasReadLock(func() error {
+		_, err := starter.PrepareFork(ctx, req)
+		return err
+	})
+}
+
+func copyGuestDirectoryWithAliasReadLock(srcDir, dstDir string) error {
+	return withSnapshotSourceAliasReadLock(func() error {
+		return forkvm.CopyGuestDirectory(srcDir, dstDir)
+	})
+}
+
+func (m *manager) copyForkSourceGuestDirectory(ctx context.Context, sourceState State, sourceID string, stored *StoredMetadata, srcDir, dstDir string) error {
+	return withSnapshotSourceAliasReadLock(func() error {
+		if sourceState == StateStandby {
+			if err := m.ensureSnapshotMemoryReady(ctx, m.paths.InstanceSnapshotLatest(sourceID), m.snapshotJobKeyForInstance(sourceID), stored.HypervisorType); err != nil {
+				return fmt.Errorf("prepare standby snapshot for fork: %w", err)
+			}
+		}
+		if err := forkvm.CopyGuestDirectory(srcDir, dstDir); err != nil {
+			if errors.Is(err, forkvm.ErrSparseCopyUnsupported) {
+				return fmt.Errorf("fork requires sparse-capable filesystem (SEEK_DATA/SEEK_HOLE unsupported): %w", err)
+			}
+			return fmt.Errorf("clone guest directory: %w", err)
+		}
+		return nil
+	})
+}
+
+func (m *manager) copySnapshotGuestDirectoryForFork(ctx context.Context, snapshotID string, hvType hypervisor.Type, dstDir string) error {
+	return withSnapshotSourceAliasReadLock(func() error {
+		if err := m.ensureSnapshotMemoryReady(ctx, m.paths.SnapshotGuestDir(snapshotID), "", hvType); err != nil {
+			return fmt.Errorf("prepare snapshot memory for fork: %w", err)
+		}
+		if err := forkvm.CopyGuestDirectory(m.paths.SnapshotGuestDir(snapshotID), dstDir); err != nil {
+			if errors.Is(err, forkvm.ErrSparseCopyUnsupported) {
+				return fmt.Errorf("fork from snapshot requires sparse-capable filesystem (SEEK_DATA/SEEK_HOLE unsupported): %w", err)
+			}
+			return fmt.Errorf("clone snapshot payload: %w", err)
+		}
+		return nil
+	})
+}
